+++ qa/lua/run.lua
@@ -135,6 +135,7 @@ local suites = {
     "test_cim_property_value_policy",
     "test_cim_tab_preview",
     "test_woc_wire_policy",
+	"test_cim_external_trait_policy",
     "test_woc_relic_policy",
 	"test_woc_blightreaper_moveset",
 	"test_woc_blightreaper_power",